cleanup evidence/list
