--- apps/app/src/app/[locale]/(app)/(dashboard)/evidence/components/data-table/index.ts
@@ -1,3 +1,0 @@
-export { DataTable } from "./EvidenceListTable";
-export { columns } from "./columns";
-export type { EvidenceTaskRow } from "./types";
