--- apps/app/src/app/[locale]/(app)/(dashboard)/evidence/components/index.ts
@@ -1,3 +1,0 @@
-export { EvidenceList } from "./EvidenceList";
-export { SkeletonTable } from "./SkeletonTable";
-export * from "./EvidenceFilters";
