+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/list/components/index.ts
@@ -1,0 +1,3 @@
+export { EvidenceList } from "./EvidenceList";
+export { SkeletonTable } from "./table/SkeletonTable";
+export * from "./table/EvidenceFilters";
