+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/list/components/table/index.ts
@@ -1,0 +1,3 @@
+export { DataTable } from "./EvidenceListTable";
+export { columns } from "./EvidenceListColumns";
+export type { EvidenceTaskRow } from "../../types";
