+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/list/hooks/useEvidenceTableContext.tsx
@@ -2,10 +2,11 @@
 
 import { createContext, useContext, useMemo, type ReactNode } from "react";
 import { useQueryState } from "nuqs";
-import { useOrganizationEvidenceTasks } from "./useEvidenceTasks";
-import type { EvidenceTaskRow } from "../components/data-table/types";
-import { ALL_DEPARTMENTS, ALL_FREQUENCIES } from "../constants";
 import type { Frequency, Departments } from "@bubba/db";
+import { useOrganizationEvidenceTasks } from "../../hooks/useEvidenceTasks";
+import { ALL_DEPARTMENTS } from "../../constants";
+import { ALL_FREQUENCIES } from "../../constants";
+import type { EvidenceTaskRow } from "../components/table";
 
 interface Assignee {
 	id: string;
