+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/list/page.tsx
@@ -1,5 +1,5 @@
-import { EvidenceList } from "../components/EvidenceList";
-import { EvidenceTableProvider } from "../hooks/useEvidenceTableContext";
+import { EvidenceList } from "./components/EvidenceList";
+import { EvidenceTableProvider } from "./hooks/useEvidenceTableContext";
 
 export default function EvidencePage() {
 	return (
